pyproject.toml
[project]
name = "aira-relay"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.26.0",
    "motor>=3.6",
    "pydantic-settings>=2.0",
    "openai>=1.30",
    "qdrant-client>=1.9",
    "pyphonetics>=0.5",
    "numpy>=1.26",
    "debugpy>=1.8",
]

[project.scripts]
aira-relay = "main:main"

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "ruff>=0.15.4",
]

[tool.ruff]
target-version = "py313"
line-length = 100
src = ["core", "relay_mcp", "."]
exclude = [".venv"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
